+++ src/java/org/apache/cassandra/cql3/QueryProcessor.java
@@ -57,7 +57,7 @@ import org.apache.cassandra.utils.*;
 
 public class QueryProcessor implements QueryHandler
 {
-    public static final CassandraVersion CQL_VERSION = new CassandraVersion("3.4.2");
+    public static final CassandraVersion CQL_VERSION = new CassandraVersion("3.4.3");
 
     public static final QueryProcessor instance = new QueryProcessor();