+++ doc/source/cql/changes.rst
@@ -26,6 +26,8 @@ The following describes the changes in each version of CQL.
 
 - Support for ``GROUP BY`` (:jira:`10707`).
 - Adds a ``DEFAULT UNSET`` option for ``INSERT JSON`` to ignore omitted columns (:jira:`11424`).
+- Allows ``null`` as a legal value for TTL on insert and update. It will be treated as equivalent to
+inserting a 0 (:jira:`12216`).
 
 3.4.2
 ^^^^^